requirements.txt
boto3==1.28.62
botocore==1.31.62
click==8.1.7
Flask==2.2.5
Flask-Cors==4.0.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
python-dateutil==2.8.2
PyYAML==6.0.1
s3transfer==0.7.0
six==1.16.0
typing_extensions==4.7.1
urllib3==1.26.17
Werkzeug==2.2.3
zipp==3.15.0
